requirements.txt
PyYAML==3.10
requests==1.2.0
wsgiref==0.1.2
